Allow opt parsing to stop on first pos arg
This is useful for commands which accept arbitrary lines which are
passed for execution further.

For example here we want `--help` to be associated with `esy build`
command:

```
% esy build --help
```

while here we want `--help` to be passed further as an arg to `dune`:

```
% esy build dune --help
```

Marking term with `~stop_on_pos:true` will make sure option parsing
stops on first positional arg.