Sort filenames before building
It's useful if building the same source code produces the same binaries,
so that everyone gets the same result.

Before, cmdliner's build system processed files in the order in which
they were returned by `Sys.readdir`, which does not return them in any
particular order. It now sorts them first.