Use mandoc if available, and UTF-8 everywhere
Reminders:
- the `-K` flag specifies the input encoding of the document
  formatter;
- the `-T` flag specifies the output encoding of the document
  formatter;

- there are two set of macros for man pages,
  - [man(7)] used mostly on Linux distros;
  - [mdoc(7)] used mostly on BSDs.
- there are two tools to render man pages, and both support man and
  mdoc:
  - GNU troff (called [groff(1)]);
  - BSD [mandoc(1)].

Both have support for automatically detecting the input encoding and
the macro set, but I've decided to specify them explicitly.

macOS ships mandoc and groff, but its groff is too old and doesn't
support the `-K` option. We have to put mandoc first to accomodate
that. It's not a problem elsewhere, because Linux distributions that
have installed mandoc by default use it by default, or otherwise mark
it as mutually exclusive with groff when groff is used by default.

[man(7)]: https://man7.org/linux/man-pages/man7/man.7.html
[mdoc(7)]: https://man.openbsd.org/mdoc.7
[groff(1)]: https://man7.org/linux/man-pages/man1/groff.1.html
[mandoc(1)]: https://man.openbsd.org/mandoc.1