Display correct subcommand in error messages
When there is an error in the command line (such as an unknown option),
this displays the command name instead of just the executable name.

This is visible in the darcs example: `./darcs_ex.exe initialize
--invalid` would previously display "darcs: unknown option '--invalid'",
but now displays "darcs initialize: ...".

This is also consistent with the `pp_try_help` path.