Add `.NOTPARALLEL` directive to Makefile
If `make` is invoked with its `-j` option, given through the command line or with the `MAKEFLAGS` environment variable, it will build the recipes in parallel.
The cmdliner build system doesn't support parallel execution an is prone to races (for instance, TOCTOU in `mkdir`).

The `.NOTPARALLEL` directive of Make disables parallel execution. The GNU Make version shipped in macOS doesn't support `.WAIT`, so for compatibility rely on this all-in-one variable.

https://www.gnu.org/software/make/manual/html_node/Parallel-Disable.html

Note that the `-j` option of Make isn't used in the opam file, where `-j%{jobs}` could have been passed.